feat: MiningBlock

+++ Princess_Cannot_Walk_Unpaved_Road/Princess_Cannot_Walk_Unpaved_Road/Source/Game/Scripts/Blocks/BlockObjectInfo.h
@@ -25,7 +25,7 @@ namespace Bisang
 
     struct MineableData
     {
-        int maxHp;
+        float maxHp;
         ToolType requiredTool;
         BlockId dropBlockId;
         int count;
